services/orchestrator/pyproject.toml
[tool.poetry]
name = "orchestrator"
version = "0.1.0"
description = "SRE Orchestrator Service"
authors = ["BMad-Agents"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.117.1"
uvicorn = "^0.37.0"
pytest = "^8.4.2"
ruff = "^0.13.1"
black = "^25.9.0"
google-generativeai = "^0.8.5"
pyyaml = "^6.0.1"

[tool.pytest.ini_options]
pythonpath = [
    "src"
]


httpx = "^0.28.1"
google-generativeai = "^0.7.0"

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
